firebase-functions/requirements.txt
functions-framework==3.*
flask==2.3.*
cors==1.0.*
PyPDF2==3.0.*
python-docx==0.8.*
pdfplumber==0.9.*
pandas==2.0.*
numpy==1.24.*
